Move catalog composables to ui package
